pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codexmcp"
version = "0.1.6"
authors = [
    { name = "Tomas Cupr", email = "work@tomcupr.com" }
]
description = "Minimal FastMCP server that wraps the OpenAI Codex CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "python-dotenv",
    "rich",
]
urls = { "Homepage" = "https://github.com/tomascupr/codexmcp" }

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=3.0.0"
]

openai = [
    "openai>=1.3.0"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["codexmcp*"]

[project.scripts]
codexmcp = "codexmcp.server:main" 

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/codexmcp"]
omit = ["tests/*"]